auth_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.1
python-jose==3.3.0
redis==5.0.1
structlog==23.2.0
